refactor: move widgets into views
