+++ app/src/main/java/com/f0x1d/logfox/extensions/views/widgets/MenuExtensions.kt
@@ -1,4 +1,4 @@
-package com.f0x1d.logfox.extensions.widgets
+package com.f0x1d.logfox.extensions.views.widgets
 
 import android.view.Menu
 import android.view.MenuItem
